refactor(cli): split core dispatch and commands
